+++ escam/dialoggreyscalesettings.cpp
@@ -47,8 +47,8 @@ void DialogGreyscaleSettings::on_spinBoxBoard_valueChanged(int value)
 {
 	mainWindow->greyscale_viewer_board = value;
 	settings.beginGroup("board" + QString::number(value));
-	uint16_t pixelcount = settings.value(settingPixelPath, settingPixelDefault).toUInt();
-	uint camcnt = settings.value(settingCamcntPath, settingCamcntDefault).toUInt();
+	uint16_t pixelcount = settings.value(settingPixelPath, settingPixelDefault).toDouble();
+	uint camcnt = settings.value(settingCamcntPath, settingCamcntDefault).toDouble();
 	settings.endGroup();
 	uint32_t nos = settings.value(settingNosPath, settingNosDefault).toDouble();
 	uint32_t block = mainWindow->ui->horizontalSliderBlock->value() - 1;
@@ -69,7 +69,7 @@ void DialogGreyscaleSettings::on_spinBoxCamera_valueChanged(int value)
 {
 	mainWindow->greyscale_viewer_camera = value;
 	settings.beginGroup("board" + QString::number(mainWindow->greyscale_viewer_board));
-	uint16_t pixelcount = settings.value(settingPixelPath, settingPixelDefault).toUInt();
+	uint16_t pixelcount = settings.value(settingPixelPath, settingPixelDefault).toDouble();
 	settings.endGroup();
 	uint32_t nos = settings.value(settingNosPath, settingNosDefault).toDouble();
 	uint32_t block = mainWindow->ui->horizontalSliderBlock->value() - 1;
