+++ escam/myqchartview.cpp
@@ -6,7 +6,7 @@ MyQChartView::MyQChartView(QWidget *parent) : QChartView(parent)
 	setRubberBand(QChartView::RectangleRubberBand);
 	QSettings settings;
 	curr_xmax = settings.value(settingPixelPath, settingPixelDefault).toReal();
-	if (settings.value(settingCameraSystemPath, settingCameraSystemDefault).toUInt() == 2)
+	if (settings.value(settingCameraSystemPath, settingCameraSystemDefault).toDouble() == 2)
 		curr_ymax = 0x3FFF;
 	else
 		curr_ymax = 0xFFFF;
